data_ingestion/requirements.txt
# Google Cloud Storage
google-cloud-storage==2.14.0
google-auth==2.25.2

# PDF Processing
PyMuPDF==1.23.8

# Embeddings
sentence-transformers==2.2.2
torch==2.1.2

# Vector Store
pymilvus==2.3.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
